+++ packages/mangler-css-classes/src/types.ts
@@ -2,44 +2,46 @@
  * The options for _WebMangler_'s built-in CSS class mangler.
  *
  * @since v0.1.0
- * @version v0.1.23
+ * @version v0.1.24
  */
 interface CssClassManglerOptions {
   /**
    * One or more patterns for CSS classes that should be mangled.
    *
-   * @default `"cls-[a-zA-Z-_]+"`
+   * @default "cls-[a-zA-Z-_]+"
    * @since v0.1.0
-   * @version v0.1.17
+   * @version v0.1.24
    */
-  classNamePattern?: string | Iterable<string>;
+  readonly classNamePattern?: string | Iterable<string>;
 
   /**
    * One or more patterns for CSS classes that should **never** be mangled.
    *
-   * @default `[]`
-   * @version v0.1.23
+   * @default []
+   * @since v0.1.23
+   * @version v0.1.24
    */
-  ignoreClassNamePattern?: string | Iterable<string>;
+  readonly ignoreClassNamePattern?: string | Iterable<string>;
 
   /**
    * A list of strings and patterns of CSS class names that should not be used.
    *
    * Patterns are supported since v0.1.7.
    *
-   * @default `[]`
+   * @default []
    * @since v0.1.0
-   * @version v0.1.17
+   * @version v0.1.24
    */
-  reservedClassNames?: Iterable<string>;
+  readonly reservedClassNames?: Iterable<string>;
 
   /**
    * A prefix to use for mangled CSS classes.
    *
-   * @default `""`
+   * @default ""
    * @since v0.1.0
+   * @version v0.1.24
    */
-  keepClassNamePrefix?: string;
+  readonly keepClassNamePrefix?: string;
 
   /**
    * A list of HTML attributes whose value to treat as `class`.
@@ -47,11 +49,11 @@ interface CssClassManglerOptions {
    * NOTE: the `class` attribute is always included and does not need to be
    * specified when using this option.
    *
-   * @default `[]`
+   * @default ["class"]
    * @since v0.1.16
-   * @version v0.1.17
+   * @version v0.1.24
    */
-  classAttributes?: Iterable<string>;
+  readonly classAttributes?: Iterable<string>;
 }
 
 export type {