+++ packages/mangler-css-classes/CHANGELOG.md
@@ -8,11 +8,12 @@ Versioning].
 
 ## [Unreleased]
 
-- _No changes yet_
+- Correct the default value for the `classAttributes` option. ([#327])
 
 ## [0.1.23] - 2021-07-25
 
 - Extract the _Webmangler Core_ CSS class mangler into this package.
 
+[#327]: https://github.com/ericcornelissen/webmangler/pull/327
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/ "Keep a CHANGELOG"
 [semantic versioning]: https://semver.org/spec/v2.0.0.html "Semantic versioning"